Dashboard Code/backend/requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-Cors==5.0.1
pyserial==3.5
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
eventlet==0.39.1
